python/requirements.txt
bip-utils
Pillow==8.3.1
numpy-stl==2.16.0
pywavefront==1.3.3
pyglet==1.5.21
